Consolidate BrowserAuthCallbackHandler and BrowserAuthorizationClient (#123)
