+++ common/changes/@itwin/browser-authorization/bdp-102-consolidate-browser-auth-callback-handler_2023-03-07-17-13.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@itwin/browser-authorization",
+      "comment": "Consolidate BrowserAuthCallbackHandler into BrowserAuthorizationClient",
+      "type": "minor"
+    }
+  ],
+  "packageName": "@itwin/browser-authorization"
+}
