+++ common/changes/@itwin/oidc-signin-tool/bdp-102-consolidate-browser-auth-callback-handler_2023-03-20-17-32.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@itwin/oidc-signin-tool",
+      "comment": "",
+      "type": "none"
+    }
+  ],
+  "packageName": "@itwin/oidc-signin-tool"
+}
