+++ packages/browser/src/index.ts
@@ -2,9 +2,9 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
-export * from "./CallbackHandler";
 export * from "./Client";
 export * from "./LoggerCategory";
+export * from "./types";
 
 /** @docs-package-description
  * The browser-authorization-client package contains classes and structures relevant to user authentication/authorization in frontend applications, with specific implementations for web browser use-cases.
